RDCC-2225: converted tests in DeleteRolesToExistingUserFuncTest class
